+++ src/khroma/windows.cljs
@@ -1,10 +1,14 @@
 (ns khroma.windows
   (:require
     [cljs.core.async :refer [chan >!]]
-    [clojure.walk :as walk])
+    [clojure.walk :as walk]
+    [khroma.util :as kutil])
   (:require-macros
     [cljs.core.async.macros :refer [go go-loop]]))
 
+
+;; Always -1, but let's expose it in case it changes
+(def none (.-WINDOW-ID-NONE js/chrome.windows))
 
 (defn get-all
   "Returns an array containing information for all open windows.
@@ -34,3 +38,24 @@
                       (>! ch (walk/keywordize-keys (js->clj window))))))
      ch)))
 
+
+(defn on-created
+  "Receives events when a window is created.
+
+  See https://developer.chrome.com/extensions/windows#event-onCreated"
+  []
+  (kutil/add-listener js/chrome.windows.onCreated :windowId))
+
+(defn on-focus-changed
+  "Receives events when a window gains or loses focus.
+
+  See https://developer.chrome.com/extensions/windows#event-onFocusChanged"
+  []
+  (kutil/add-listener js/chrome.windows.onFocusChanged :windowId))
+
+(defn on-removed
+  "Receives events when a window is removed
+
+  See https://developer.chrome.com/extensions/windows#event-onRemoved"
+  []
+  (kutil/add-listener js/chrome.windows.onRemoved :windowId))