+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `identity`'s namespace was mistakenly `extension`
 - New `tabs/remove`, `tabs/query`
 - New `windows/get-all`
+- New event handlers for `windows/on-created`, `windows/on-removed`, `windows/on-focus-changed`
 
 ## 0.1.0